feat(Interface): 全站卡片编辑交互统一为右侧 Drawer 抽屉，替代居中模态框;
- SkillFormDialog → SkillFormDrawer：OverlayDismissLayer 替换为 BaseDrawer，新增脏检确认
- ToolFormDialog → ToolFormDrawer：同上，保留非对称 footer（Test Connection + Cancel/Submit），多状态脏检
- McpServerFormDialog → McpServerFormDrawer：同上，条件传输区段不受影响
- RoutineEditDrawer：移除手动 useFocusTrap + translateX CSS 动画 + overlay/panel DOM，改用 BaseDrawer
- BaseDrawer：新增 headerActions 插槽，支持 RoutineEditDrawer 的 Full View 链接
- 更新 SkillsLayout、McpLayout 测试 mock 路径/组件名
