--- apps/negentropy-ui/app/interface/mcp/_components/McpServerFormDialog.tsx
@@ -1,365 +1,0 @@
-/* eslint-disable react-hooks/set-state-in-effect --
- * React 19 + eslint-plugin-react-hooks v7.1.1 的 React Compiler 兼容新规则集
- * 在该文件中命中既有代码模式（useEffect 内调用 fetcher / ref 写入 / deps 校验等）。
- * 这些代码功能正确，仅是新规则严格度提升导致的告警；
- * TODO(react-compiler): 按 React Compiler 范式 / SWR / useSyncExternalStore 重构。
- */
-"use client";
-
-import { useState, useEffect } from "react";
-import { OverlayDismissLayer } from "@/components/ui/OverlayDismissLayer";
-import { Button } from "@/components/ui/Button";
-
-interface McpServer {
-  id: string;
-  name: string;
-  display_name: string | null;
-  description: string | null;
-  transport_type: string;
-  command: string | null;
-  args: string[];
-  env: Record<string, string>;
-  url: string | null;
-  headers: Record<string, string>;
-  is_enabled: boolean;
-  auto_start: boolean;
-  config: Record<string, unknown>;
-  visibility: string;
-}
-
-interface McpServerFormDialogProps {
-  open: boolean;
-  onClose: () => void;
-  onSubmit: (data: Record<string, unknown>) => Promise<void>;
-  server: McpServer | null;
-}
-
-export function McpServerFormDialog({
-  open,
-  onClose,
-  onSubmit,
-  server,
-}: McpServerFormDialogProps) {
-  const [formData, setFormData] = useState({
-    name: "",
-    display_name: "",
-    description: "",
-    transport_type: "stdio",
-    command: "",
-    args: "",
-    env: "",
-    url: "",
-    headers: "",
-    is_enabled: true,
-    auto_start: false,
-    visibility: "private",
-  });
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (server) {
-      setFormData({
-        name: server.name,
-        display_name: server.display_name || "",
-        description: server.description || "",
-        transport_type: server.transport_type,
-        command: server.command || "",
-        args: Array.isArray(server.args) ? server.args.join("\n") : "",
-        env: typeof server.env === "object" ? JSON.stringify(server.env, null, 2) : "{}",
-        url: server.url || "",
-        headers: typeof server.headers === "object" ? JSON.stringify(server.headers, null, 2) : "{}",
-        is_enabled: server.is_enabled,
-        auto_start: server.auto_start,
-        visibility: server.visibility,
-      });
-    } else {
-      setFormData({
-        name: "",
-        display_name: "",
-        description: "",
-        transport_type: "stdio",
-        command: "",
-        args: "",
-        env: "{}",
-        url: "",
-        headers: "{}",
-        is_enabled: true,
-        auto_start: false,
-        visibility: "private",
-      });
-    }
-    setError(null);
-  }, [server, open]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-    setError(null);
-
-    try {
-      const data: Record<string, unknown> = {
-        name: formData.name,
-        display_name: formData.display_name || null,
-        description: formData.description || null,
-        transport_type: formData.transport_type,
-        is_enabled: formData.is_enabled,
-        auto_start: formData.auto_start,
-        visibility: formData.visibility,
-      };
-
-      if (formData.transport_type === "stdio") {
-        data.command = formData.command || null;
-        data.args = formData.args
-          .split("\n")
-          .map((s) => s.trim())
-          .filter(Boolean);
-        try {
-          data.env = JSON.parse(formData.env || "{}");
-        } catch {
-          throw new Error("Invalid JSON in environment variables");
-        }
-      } else {
-        data.url = formData.url || null;
-        try {
-          data.headers = JSON.parse(formData.headers || "{}");
-        } catch {
-          throw new Error("Invalid JSON in headers");
-        }
-      }
-
-      await onSubmit(data);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "An error occurred");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  if (!open) return null;
-
-  return (
-    <OverlayDismissLayer
-      open={open}
-      onClose={onClose}
-      busy={loading}
-
-      containerClassName="flex min-h-full items-start justify-center overflow-y-auto p-3 sm:p-6"
-      contentClassName="my-3 flex max-h-[calc(100vh-1rem)] w-full max-w-6xl flex-col overflow-hidden rounded-modal border border-border bg-card shadow-xl sm:max-h-[calc(100vh-2rem)]"
-    >
-          <div className="border-b border-border px-5 py-4 sm:px-6">
-            <h2 className="text-lg font-semibold text-foreground">
-              {server ? "Edit MCP Server" : "Add MCP Server"}
-            </h2>
-            <p className="mt-1 text-sm text-text-muted">
-              Configure transport and runtime options with a consistent, scannable interface form layout.
-            </p>
-          </div>
-
-          <form onSubmit={handleSubmit} className="flex min-h-0 flex-1 flex-col">
-            <div className="min-h-0 flex-1 space-y-6 overflow-y-auto px-5 py-5 sm:px-6">
-              {error && (
-                <div className="rounded-md bg-red-50 p-3 text-sm text-red-600 dark:bg-red-900/20 dark:text-red-400">
-                  {error}
-                </div>
-              )}
-
-              <section className="space-y-4">
-                <h3 className="text-xs font-semibold uppercase tracking-wide text-text-muted">
-                  Basic Information
-                </h3>
-                <div className="grid gap-4 lg:grid-cols-2">
-                  <div>
-                    <label className="mb-1 block text-sm font-medium text-text-secondary">
-                      Name *
-                    </label>
-                    <input
-                      type="text"
-                      value={formData.name}
-                      onChange={(e) => setFormData({ ...formData, name: e.target.value })}
-                      className="w-full rounded-md border border-border bg-input px-3 py-2 text-sm text-foreground"
-                      placeholder="my-mcp-server"
-                      required
-                    />
-                  </div>
-                  <div>
-                    <label className="mb-1 block text-sm font-medium text-text-secondary">
-                      Display Name
-                    </label>
-                    <input
-                      type="text"
-                      value={formData.display_name}
-                      onChange={(e) => setFormData({ ...formData, display_name: e.target.value })}
-                      className="w-full rounded-md border border-border bg-input px-3 py-2 text-sm text-foreground"
-                      placeholder="My MCP Server"
-                    />
-                  </div>
-                  <div className="lg:col-span-2">
-                    <label className="mb-1 block text-sm font-medium text-text-secondary">
-                      Description
-                    </label>
-                    <textarea
-                      value={formData.description}
-                      onChange={(e) => setFormData({ ...formData, description: e.target.value })}
-                      className="w-full rounded-md border border-border bg-input px-3 py-2 text-sm text-foreground"
-                      rows={2}
-                      placeholder="Description of this MCP server"
-                    />
-                  </div>
-                </div>
-              </section>
-
-              <section className="space-y-4">
-                <h3 className="text-xs font-semibold uppercase tracking-wide text-text-muted">
-                  Runtime Setup
-                </h3>
-                <div className="grid gap-4 sm:grid-cols-2 xl:grid-cols-4">
-                  <div>
-                    <label className="mb-1 block text-sm font-medium text-text-secondary">
-                      Transport Type *
-                    </label>
-                    <select
-                      value={formData.transport_type}
-                      onChange={(e) => setFormData({ ...formData, transport_type: e.target.value })}
-                      className="w-full rounded-md border border-border bg-input px-3 py-2 text-sm text-foreground"
-                    >
-                      <option value="stdio">STDIO</option>
-                      <option value="http">HTTP (Streamable)</option>
-                      <option value="sse">SSE</option>
-                    </select>
-                  </div>
-                  <div>
-                    <label className="mb-1 block text-sm font-medium text-text-secondary">
-                      Visibility
-                    </label>
-                    <select
-                      value={formData.visibility}
-                      onChange={(e) => setFormData({ ...formData, visibility: e.target.value })}
-                      className="w-full rounded-md border border-border bg-input px-3 py-2 text-sm text-foreground"
-                    >
-                      <option value="private">Private</option>
-                      <option value="shared">Shared</option>
-                      <option value="public">Public</option>
-                    </select>
-                  </div>
-                  <div className="flex items-end">
-                    <label className="flex w-full items-center gap-2 rounded-md border border-border px-3 py-2 text-sm text-text-secondary">
-                      <input
-                        type="checkbox"
-                        checked={formData.is_enabled}
-                        onChange={(e) => setFormData({ ...formData, is_enabled: e.target.checked })}
-                        className="rounded border-border"
-                      />
-                      Enabled
-                    </label>
-                  </div>
-                  <div className="flex items-end">
-                    <label className="flex w-full items-center gap-2 rounded-md border border-border px-3 py-2 text-sm text-text-secondary">
-                      <input
-                        type="checkbox"
-                        checked={formData.auto_start}
-                        onChange={(e) => setFormData({ ...formData, auto_start: e.target.checked })}
-                        className="rounded border-border"
-                      />
-                      Auto-start
-                    </label>
-                  </div>
-                </div>
-              </section>
-
-              <section className="space-y-4">
-                <h3 className="text-xs font-semibold uppercase tracking-wide text-text-muted">
-                  Connection Details
-                </h3>
-
-                {formData.transport_type === "stdio" ? (
-                  <div className="grid gap-4 lg:grid-cols-2">
-                    <div className="lg:col-span-2">
-                      <label className="mb-1 block text-sm font-medium text-text-secondary">
-                        Command *
-                      </label>
-                      <input
-                        type="text"
-                        value={formData.command}
-                        onChange={(e) => setFormData({ ...formData, command: e.target.value })}
-                        className="w-full rounded-md border border-border bg-input px-3 py-2 text-sm font-mono text-foreground"
-                        placeholder="npx"
-                      />
-                    </div>
-                    <div>
-                      <label className="mb-1 block text-sm font-medium text-text-secondary">
-                        Arguments (one per line)
-                      </label>
-                      <textarea
-                        value={formData.args}
-                        onChange={(e) => setFormData({ ...formData, args: e.target.value })}
-                        className="min-h-[200px] w-full rounded-md border border-border bg-input px-3 py-2 text-sm font-mono text-foreground"
-                        rows={7}
-                        placeholder="-y&#10;@modelcontextprotocol/server-filesystem&#10;/path/to/allowed/dir"
-                      />
-                    </div>
-                    <div>
-                      <label className="mb-1 block text-sm font-medium text-text-secondary">
-                        Environment Variables (JSON)
-                      </label>
-                      <textarea
-                        value={formData.env}
-                        onChange={(e) => setFormData({ ...formData, env: e.target.value })}
-                        className="min-h-[200px] w-full rounded-md border border-border bg-input px-3 py-2 text-sm font-mono text-foreground"
-                        rows={7}
-                        placeholder='{"API_KEY": "xxx"}'
-                      />
-                    </div>
-                  </div>
-                ) : (
-                  <div className="grid gap-4 lg:grid-cols-2">
-                    <div>
-                      <label className="mb-1 block text-sm font-medium text-text-secondary">
-                        URL *
-                      </label>
-                      <input
-                        type="url"
-                        value={formData.url}
-                        onChange={(e) => setFormData({ ...formData, url: e.target.value })}
-                        className="w-full rounded-md border border-border bg-input px-3 py-2 text-sm text-foreground"
-                        placeholder={formData.transport_type === "http"
-                          ? "http://localhost:8080/mcp"
-                          : "http://localhost:8080/sse"}
-                      />
-                    </div>
-                    <div>
-                      <label className="mb-1 block text-sm font-medium text-text-secondary">
-                        Headers (JSON)
-                      </label>
-                      <textarea
-                        value={formData.headers}
-                        onChange={(e) => setFormData({ ...formData, headers: e.target.value })}
-                        className="min-h-[200px] w-full rounded-md border border-border bg-input px-3 py-2 text-sm font-mono text-foreground"
-                        rows={7}
-                        placeholder='{"Authorization": "Bearer xxx"}'
-                      />
-                    </div>
-                  </div>
-                )}
-              </section>
-            </div>
-
-            <div className="flex shrink-0 justify-end gap-3 border-t border-border bg-card px-5 py-4 sm:px-6">
-              <Button
-                variant="ghost"
-                onClick={onClose}
-              >
-                Cancel
-              </Button>
-              <Button
-                variant="neutral"
-                type="submit"
-                disabled={loading}
-              >
-                {loading ? "Saving..." : server ? "Update" : "Create"}
-              </Button>
-            </div>
-          </form>
-    </OverlayDismissLayer>
-  );
-}
